Fixing ARM platforms and adding port for the Retrostone 2. Bittboy build now also use the ARM backend as well. Fix a bug with the crc32.c code as it should be placed before the other one.

+++ crc32.h
@@ -1,5 +1,5 @@
 #ifndef CRC32_H
 #define CRC32_H
-int Crc32_ComputeFile( FILE *file, unsigned long *outCrc32 );
-unsigned long Crc32_ComputeBuf( unsigned long inCrc32, const void *buf, size_t bufLen );
+int Crc32_ComputeFile( FILE *file, u32 *outCrc32 );
+u32 Crc32_ComputeBuf( u32 inCrc32, const void *buf, size_t bufLen );
 #endif
